backend/requirements-dev.txt
python-dotenv>=1.2.1
logfire>=4.25.0
boto3>=1.42.52
pandas==3.0.1
pydantic-ai-slim[bedrock]==1.62.0
aws-lambda-powertools>=3.24.0
pylint>=4.0.4
black>=26.1.0